% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{Birthday_fun}
\alias{Birthday_fun}
\title{Title Birthday_fun}
\usage{
Birthday_fun(Iter = 50, size = NULL, plot = TRUE)
}
\arguments{
\item{Iter}{Input numeric: This is the number of iterations the function will run through.
Note: The function can be quite slow if Iter >= 1000. However, 200 is more
then enough to show convergence.}

\item{size}{Input numeric vector: This is a vector containing the different
people group sizes you would like. Default is c(10,20,30,40)
Note: Values below 50 work best.}

\item{plot}{Input logical: Setting this to FALSE will suppress the graphing.}
}
\value{
Output data.frame of each size value and their running probabilities.
}
\description{
Title Birthday_fun
}
\examples{

Birthday_fun()

Birthday_fun(Iter = 100, size = c(1,5,10,15,20))

}
