% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{blackjack_bust}
\alias{blackjack_bust}
\title{Title blackjack_bust}
\usage{
blackjack_bust()
}
\value{
list of c(time_to_bust, data)
        time_to_bust = List of draws need to "bust"
        data = list of "time_to_bust", Frequency of the number of cards
        at "bust", and probabilities.
}
\description{
This is an Input Driven Black-Jack Bust Simulation version of
 blackjack_bust_auto()
 Note: at the moment this function does not produce graphics automatically.
}
\examples{

blackjack_bust()

}
