% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_qq.R
\name{build_QQ}
\alias{build_QQ}
\title{Title build_QQ}
\usage{
build_QQ(data, vs_dist = "normal", alpha = 1, beta = 1)
}
\arguments{
\item{data}{Input Vector of numeric data.}

\item{vs_dist}{Input String being one of:
"normal","gamma","uniform","poisson","binomial","geometric","chi-squared"}

\item{alpha}{Input Numeric:
This is the default parameter for the above distributions.}

\item{beta}{Input Numeric:
This is a slot for the second parameter for the above distributions.}
}
\value{
QQ-Plot and density plot of the data.
}
\description{
Title build_QQ
}
\examples{
data <- sample(1:1000,100)
build_QQ(data)

build_QQ(data, vs_dist = "gamma", alpha = 2, beta = 4)

build_QQ(data, vs_dist = "poisson", alpha = 10)
}
