% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{card_count}
\alias{card_count}
\title{Title card_count()
      This function will calculate card values and is needed for other simulation
      functions in the APStatTools package.}
\usage{
card_count(draw, draw_values = 0)
}
\arguments{
\item{draw}{Input String: A list of drawn card values
being "2","3","4","5","6","7","8","9","A","J","K","Q"}

\item{draw_values}{Input numeric: This is for a previously defined total card values.
You will usually leave this as 0}
}
\value{
Output list of c((draw, draw_values, value))
        draw = your cards
        draw_values = each card's worth
        value = the total worth.
}
\description{
Title card_count()
      This function will calculate card values and is needed for other simulation
      functions in the APStatTools package.
}
\examples{

card_count(draw = c("2","3","4"))
}
