% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{chi_squared.gof}
\alias{chi_squared.gof}
\title{Title chi_squared.gof}
\usage{
chi_squared.gof(
  obs_table,
  expected_table = NULL,
  expected_as_count = FALSE,
  row_totals = FALSE,
  graph = TRUE
)
}
\arguments{
\item{obs_table}{Input Numeric Vector: This is the observed vector of categorical values.
Note: If the total is included please set row_totals = TRUE}

\item{expected_table}{Input Numeric Vector: This is a vector of either expected
values or proportions to be computed in Pearson's Chi-Squared test.
If the values are counts and not proportions, please set expected_as_count = TRUE.}

\item{expected_as_count}{Input Logical. If expected_table is of proportions let this be FALSE.
If the expected_table is of counts let this be TRUE.}

\item{row_totals}{Input Logical: If the obs_table includes the total, let this be TRUE.
Note: It is defaulted to be FALSE, and will include the total in the calculation
if a total is actually included. This will lead to errors. Also, you will want to ensure
that both the obs_table and expected_table either both have totals or both do not. The
function will either remove the last element of both vectors, or neither vectors.}

\item{graph}{Input Logical: Set this to FALSE if you want to suppress graphing the density function.}
}
\value{
Output: This will return a data.frame of statistical values and a graphic if desired.
        the data frame will contain some of the following values, given the appropriate test.
        Test Name, p_value, test_statistic, obs_table, expected_count,
        and chi_squared_values test statistic values.
}
\description{
Title chi_squared.gof
}
\examples{

X <- c(10,13,14,20,16,73)
Y <- c(8,12,8,16,13,57)
chi_squared.gof(obs_table = X, expected_table = Y, expected_as_count = TRUE,
                row_totals = TRUE, graph = TRUE)

}
