% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{chi_squared.ind}
\alias{chi_squared.ind}
\title{Title chi_squared.ind   Chi-Squared Test for Homogeneity or Independence.}
\usage{
chi_squared.ind(obs_table, mat_totals = FALSE, graph = TRUE)
}
\arguments{
\item{obs_table}{Input Numeric Vector: This is the observed vector of categorical values.
Note: If the total is included please set row_totals = TRUE}

\item{mat_totals}{Input Logical: If FALSE this function will generate the total columns and rows for the user.
However, it will add both the row and column totals. It is not recommended to use this feature
to add just one of the row's or column's totals, as you will end up with an additional unwanted set of totals.}

\item{graph}{Input Logical: Set this to FALSE if you want to suppress graphing the density function.}
}
\value{
Output: This will return a data.frame of statistical values and a graphic if desired.
        the data frame will contain some of the following values, given the appropriate test.
        Test Name, p_value, test_statistic, obs_table, expected_count,
        and chi_squared_values test statistic values.
}
\description{
Title chi_squared.ind   Chi-Squared Test for Homogeneity or Independence.
}
\examples{

X <- matrix(rnorm(20,10,3), nrow = 4)
chi_squared.ind(obs_table = X, mat_totals = FALSE, graph = TRUE)

}
