% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{confidence_interval}
\alias{confidence_interval}
\title{Title confidence_interval}
\usage{
confidence_interval(mean = 0, sigma = 1, n = 20, level = 0.95, trials = 25)
}
\arguments{
\item{mean}{Input numeric: The mean value of your prior.}

\item{sigma}{Input numeric: The standard deviation of your prior.}

\item{n}{Input numeric: The sample size you would like to take.}

\item{level}{Input numeric: The confidence level between 0 and 1.}

\item{trials}{Input numeric. Number of times you would like to run the graph.}
}
\value{
Plot
}
\description{
This is designed to build a confidence interval plot.
}
\examples{

confidence_interval(mean = 0, sigma = 1, n = 20, level = .95, trials = 20)
confidence_interval(mean = 10, sigma = 3, n = 50, level = .90, trials = 30)

}
