% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_probs.R
\name{find_probs}
\alias{find_probs}
\title{Title find_probs()}
\usage{
find_probs(
  bound,
  type = "normal",
  tail = "left",
  mean = 0,
  sd = 1,
  df = 1,
  prob = 0.5,
  trials = 10,
  inverse = FALSE
)
}
\arguments{
\item{bound}{Vector of length one or two: c(left,right). This is to define the boundaries of the shaded region of the graph.
Notes: "inner" and "two" require both the left and right input.}

\item{type}{String being any of "normal", "t-dist", "chi-squared", "binomial"}

\item{tail}{String being any of "left", "right", "inner", "outer" or "two", "left_not_equal", "right_not_equal"
Notes: "left_not_equal" and "left_not_equal" are intended for the "binomial" only.
Notes: "out" and "two" will produce the same output.}

\item{mean}{Numeric: vector of one value. The mean of the distribution}

\item{sd}{Numeric: vector of one value. The standard deviation of a distribution.}

\item{df}{Numeric: vector of one value. The degree of freedom of a distribution.}

\item{prob}{Numeric: vector of one value. The probability of success in the binomial distribution.}

\item{trials}{Numeric: vector of one value. The number of trials in the binomial distribution.}

\item{inverse}{Logical: If true the function will assume bound is a Area, and the function will attempt to find the
quantile that produces a probability of "bound"
Note: This function assumes only "left" as a tail. Imputing other tails will have no effect on the output.}
}
\value{
Numeric: Probability = being the calculated probability of the region supplied. (The cdf)
        Numeric: Value = being the calculated value that produces the area supplied in bound. (the quantile)
}
\description{
Description: This is a simple wrapper function for finding the probabilities for various distributions.
}
