% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Title}
\usage{
get_data(data_name = NULL, ap_data_set = TRUE)
}
\arguments{
\item{data_name}{}
}
\description{
Title
}
\examples{



Note: This function will allow you to enter you data set name as
}
