% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measures.R
\name{get_measures}
\alias{get_measures}
\title{Title get_measures()}
\usage{
get_measures(
  input = FALSE,
  X = NULL,
  Y = NULL,
  freq = c(1),
  frequency = FALSE,
  regression = FALSE,
  outliers = FALSE
)
}
\arguments{
\item{input}{logical. If TRUE it will prompt the user to input their data into a spreadsheet.
If FALSE the function will expect the user to input the data into X and/or Y.}

\item{X}{Vector of numeric data. This is not intended to be used for a Matrix. Please use input = TRUE to input a matrix.}

\item{Y}{Vector of numeric data}

\item{regression}{logical. IF TRUE the function will print out the summary(lm) for Y ~ X.}

\item{outliers}{String being any of "FALSE", "1.5 IQR", "3sd Rule".
If FALSE the function will not print bounds for outliers. If a rule is implemented the function will
print a lower and upper boundary for outliers of the given rule.}
}
\value{
list($data, $measures, $LSRL)
        data: Matrix that contains the data the user input.
        measures: is a data.frame() containing statistical measures about the users data.
           measures can contain any of list("data","totals","n","mean","pop_sd","sample_sd","Q1","Q2","Q3","IQR","lower_outliers","upper_outliers")
         LSRL: the list created by summary(lm()) in base R
}
\description{
Description: This function will take user input data or data sets and compute a large number of statistical values.
}
