% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{qq_dem}
\alias{qq_dem}
\title{Title qq_dem}
\usage{
qq_dem()
}
\value{
Plot:
        This plot will build a lot of QQ plots vs normal plots, and histograms of the simulated data.
        In Rstudio you can click through the arrows to look through various examples. In R, click on the graph to see the next on.
}
\description{
This is to demonstrate the effectiveness of the QQ plots against various distributions.
}
\examples{

qq_dem()


}
