% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{t_test.one}
\alias{t_test.one}
\title{Title t_test.one   One sample t-test for means.}
\usage{
t_test.one(null, x_bar, sigma, n, tail = "left", graph = TRUE)
}
\arguments{
\item{x_bar}{Input Vector: This is the value of the sample means. It assumes the standard error adjustment has not been taken.}

\item{sigma}{Input Vector: This is the value of the sample standard deviations.}

\item{n}{Input Vector: This is the value of the sample sizes.}

\item{tail}{Input String: Being one of "left" or "right".
Note: this function will assume two tailed null hypothesis if anything other
then "left" or "right" is used.}

\item{graph}{Input Logical: Set this to FALSE if you want to suppress graphing the density function.}

\item{nullInput}{Numeric: This is a single value for the Null Hypothesis's expected value.}
}
\value{
Output: This will return a data.frame of statistical values and a graphic if desired.
        the data frame will contain Test Name, p_value, test_statistic, n, standard_error, df,
}
\description{
Title t_test.one   One sample t-test for means.
}
\examples{
t_test.one(null = 12, x_bar = 10, sigma = 5, n = 25, tail = "left", graph = TRUE)

}
