% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{t_test.paired}
\alias{t_test.paired}
\title{Title t_test.paired   Pooled two sample t-test}
\usage{
t_test.paired(null = 0, table_one, table_two, tail = "two", graph = TRUE)
}
\arguments{
\item{null}{Input Numeric: This is a single value for the Null Hypothesis's expected value.}

\item{table_one}{Input Vector: This is a vector of numeric values intended for the
t_test.paired test only. This is the first vector to be compared. The function will assume table_one - table_two.}

\item{table_two}{Input Vector: This is a vector of numeric values intended for the
t_test.paired test only. This is the second vector to be compared. The function will assume table_one - table_two.}

\item{tail}{Input String: Being one of "left" or "right".
Note: this function will assume two tailed null hypothesis if anything other
then "left" or "right" is used.}

\item{graph}{Input Logical: Set this to FALSE if you want to suppress graphing the density function.}
}
\value{
Output: This will return a data.frame of statistical values and a graphic if desired.
        the data frame will contain Test Name, p_value, test_statistic, n, standard_error, df,
}
\description{
Note: This will assume the expected difference is calculated as table_one - table_two.
   Please input the data in the order you would like them subtracted, to match the above.
}
\examples{
X <- c(10,11,12,13,14,15,16,17,18,19,20)
Y <- c(12,11,13,14,14,9,15,13,20,14,18)
t_test.paired(null = 0, table_one = X, table_two = Y, tail = "left", graph = TRUE)

}
