% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{t_test.two}
\alias{t_test.two}
\title{Title t_test.two   Two sample t-test on means}
\usage{
t_test.two(null = 0, x_bar, sigma, n, tail = "left", graph = TRUE)
}
\arguments{
\item{x_bar}{Input Vector of two: This is the value of the sample means c(left, right). It assumes the standard error adjustment has not been taken.}

\item{sigma}{Input Vector of two: This is the value of the sample standard deviations c(left, right).}

\item{n}{Input Vector f two: This is the value of the sample sizes c(left, right).}

\item{tail}{Input String: Being one of "left" or "right".
Note: this function will assume two tailed null hypothesis if anything other
then "left" or "right" is used.}

\item{graph}{Input Logical: Set this to FALSE if you want to suppress graphing the density function.}

\item{nullInput}{Numeric: This is a single value for the Null Hypothesis's expected value.}
}
\value{
Output: This will return a data.frame of statistical values and a graphic if desired.
        the data frame will contain Test Name, p_value, test_statistic, n, standard_error, df,
}
\description{
Title t_test.two   Two sample t-test on means
}
\examples{
t_test.two(null = 0, x_bar = c(11,10), sigma = c(4,5), n = c(30,25), tail = "left", graph = TRUE)


}
