% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{z_test.one}
\alias{z_test.one}
\title{Title z_test.one  one sample proportion test.}
\usage{
z_test.one(null, p_hat, n, tail = "left", graph = TRUE)
}
\arguments{
\item{null}{Input Numeric: This is a single value for the Null Hypothesis's expected value.}

\item{p_hat}{Input Vector: This is the value of the sample proportions.
It assumes the standard error adjustment has not been taken, and that the value is in decimal form.}

\item{n}{Input Vector: This is the value of the sample sizes.}

\item{tail}{Input String: Being one of "left" or "right".
Note: this function will assume two tailed null hypothesis if anything other
then "left" or "right" is used. Be aware that the function will also assume a "left" tail when preforming
the two tailed calculation. To avoid this, set tail to "right" and multiply by 2.}

\item{graph}{Input Logical: Set this to FALSE if you want to suppress graphing the density function.}
}
\value{
Output: This will return a data.frame of statistical values and a graphic if desired.
        the data frame will contain some of the following values, given the appropriate test.
        Test Name, p_value, test_statistic, n, standard_error.
}
\description{
Note: This test will fail if null = 0
}
\examples{

z_test.one(null = .4, p_hat = .6, n = 10, tail = "left", graph = TRUE)


}
