% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{z_test.pooled}
\alias{z_test.pooled}
\title{Title z_test.pooled    Two sample test on proportions.}
\usage{
z_test.pooled(p_hat, n, tail = "left", graph = TRUE)
}
\arguments{
\item{p_hat}{Input Vector of two: This is the value of the sample proportions c(left,right)
It assumes the standard error adjustment has not been taken, and that the value is in decimal form.}

\item{n}{Input Vector of two: This is the value of the sample sizes  c(left,right).}

\item{tail}{Input String: Being one of "left" or "right".
Note: this function will assume two tailed null hypothesis if anything other
then "left" or "right" is used. Be aware that the function will also assume a "left" tail when preforming
the two tailed calculation. To avoid this, set tail to "right" and multiply by 2.}

\item{graph}{Input Logical: Set this to FALSE if you want to suppress graphing the density function.}
}
\value{
Output: This will return a data.frame of statistical values and a graphic if desired.
        the data frame will contain some of the following values, given the appropriate test.
        Test Name, p_value, test_statistic, n, standard_error.
}
\description{
Title z_test.pooled    Two sample test on proportions.
}
\examples{

z_test.pooled(p_hat = c(.4,.5), n = c(25,24), tail = "left", graph = TRUE)


}
