% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measures.R
\name{get_measures}
\alias{get_measures}
\title{Title get_measures()}
\usage{
get_measures(
  X = NULL,
  input = FALSE,
  Y = NULL,
  regression = FALSE,
  outliers = FALSE,
  names = NULL
)
}
\arguments{
\item{X}{Input Vector of numeric data.}

\item{input}{Input logical. If TRUE it will prompt the user to input their data into a spreadsheet.
If FALSE the function will expect the user to input the data into X and/or Y.}

\item{Y}{Inopt Vector of numeric data.}

\item{regression}{Input logical. IF TRUE the function will print out the summary(lm) for Y ~ X.}

\item{outliers}{Input String being any of "FALSE", "1.5 IQR", "3sd Rule".
If FALSE the function will not print bounds for outliers. If a rule is implemented the function will
print a lower and upper boundary for outliers of the given rule.}

\item{names}{Input String. This is a vector of strings if you wish to provide your data with
specific column names. If you used input = TRUE, please leave this NULL, and input your
names into the prompted window.}
}
\value{
list($data, $measures, $LSRL)
        data: Matrix that contains the data the user input.
        measures: is a data.frame() containing statistical measures about the users data.
           measures can contain any of list("data","totals","n","mean","pop_sd","sample_sd","Q1","Q2","Q3","IQR","lower_outliers","upper_outliers")
         LSRL: the list created by summary(lm()) in base R
}
\description{
Description: This function will take user input data or data sets and compute a large number of statistical values.
}
\examples{
# This will prompt you to input a set of data.
get_measures(input=TRUE)

data <- c(sample(1:100,15))
get_measures(X = data)

# This will allow for outliers to be printed, and will assing the name "BOB" to the data.
get_measures(X = data, outliers = TRUE, names = "BOB")

# Two variate data
data <- c(sample(1:100,15))
data2 <- seq(1:15)
get_measures(X = data, Y = data2, outliers = TRUE, names = "BOB")
get_measures(X = data, Y = data2, regression = TRUE, outliers = TRUE, names = "BOB")

data <- matrix(sample(1:100, 50), ncol <- 25)
get_measures(X = data, outliers = TRUE, names = c("BOB","JILL"))


}
